requirements.txt
click==5.1
python-dotenv==0.1.3
requests==2.7.0
slacker==0.7.3
python-forecastio==1.3.4
python-daemon==2.0.6
PyYAML==3.11
websocket-client==0.32.0
slackclient==0.16